move Framework tests to separate folder, refactors
