+++ tests.Framework/Lussatite.FeatureManagement.Net48.Tests/LussatiteFeatureManagerTests.cs
@@ -1,7 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Lussatite.FeatureManagement.SessionManagers.Framework;
-using Lussatite.FeatureManagement.TestingCommon;
 using Xunit;
 
 namespace Lussatite.FeatureManagement.Net48.Tests
