+++ tests.Framework/Lussatite.FeatureManagement.Net48.Tests/Properties/AssemblyInfo.cs
