+++ tests.Framework/Lussatite.FeatureManagement.Net48.Tests/app.config
