+++ tests.Framework/Lussatite.FeatureManagement.Net48.Tests/packages.config
