+++ tests.Framework/Lussatite.FeatureManagement.SessionManagers.Framework.Tests/Properties/AssemblyInfo.cs
