+++ tests.Framework/Lussatite.FeatureManagement.SessionManagers.Framework.Tests/TestFeatures.cs
