+++ tests.Framework/Lussatite.FeatureManagement.SessionManagers.Framework.Tests/app.config
