+++ tests.Framework/Lussatite.FeatureManagement.SessionManagers.Framework.Tests/packages.config
