+++ tests/Lussatite.FeatureManagement.NetCore31.Tests/LussatiteFeatureManagerTests.cs
@@ -1,7 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Lussatite.FeatureManagement.SessionManagers;
-using Lussatite.FeatureManagement.TestingCommon;
 using Microsoft.Extensions.Configuration;
 using Xunit;
