--- tests/Lussatite.FeatureManagement.TestingCommon/Lussatite.FeatureManagement.TestingCommon.csproj
@@ -1,9 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <WarningsAsErrors>true</WarningsAsErrors>
-    <LangVersion>10</LangVersion>
-  </PropertyGroup>
-
-</Project>
